Adds block mining

+++ include/block.h
@@ -24,7 +24,7 @@
 typedef struct block_t {
     time_t created_at;
     linked_list_t *transaction_list;
-    uint32_t proof_of_work;
+    uint64_t proof_of_work;
     sha_256_t previous_block_hash;
 } block_t;
 
@@ -41,7 +41,7 @@ typedef struct block_t {
 return_code_t block_create(
     block_t **block,
     linked_list_t *transaction_list,
-    uint32_t proof_of_work,
+    uint64_t proof_of_work,
     sha_256_t previous_block_hash
 );
